requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
psycopg2-binary>=2.9.6
httpx>=0.24.0
pydantic[email]>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
aiofiles>=23.0.0
python-dotenv>=1.0.0
unidecode>=1.3.0
# FIX 15: Updated PyPDF2 to pypdf (more secure and maintained)
pypdf>=4.0.0
python-docx>=0.8.11
# CrawlKit for web crawling
crawlkit>=0.1.0
# Multi-LLM Provider support
anthropic>=0.25.0
openai>=1.0.0
cryptography>=42.0.0
google-api-python-client>=2.170.0
google-auth>=2.40.0
google-auth-httplib2>=0.2.0
# HuggingFace data integration
datasets>=2.20.0
beautifulsoup4>=4.12.0
# Embeddings for chunking (Lucky will use this)
sentence-transformers>=3.0.0
